added overlay_enabled array
main_window now takes the overlay_enabled array as a parameter.
Fix bug where nullptr is not a valid nitialisation for cv::String but would only cause a runtime fault.
Alter GUI dimensions of the Settings window and added column for checkboxes that allow user to turn overlay on and off.
Leave room in the Settings window to allow future addition of  trackbars for alpha values