added height and width to opencv_window
Remember the height and width of the opened video input (webcam), in preparation for overlay of a fixed image (JPEG) that will need to be scaled and centered so that both matrices have the same dimensions before alpha blending.
take_picture now throws a domain_error if the captured picture from a live webcam does not have the same dimension as the opened webcam.  This exception is fatal and is not caught anywhere, causing the application to abort by design.
opencv_ui remembers all extra window dimensions as a collection of pairs of double